openai_demo.py
import openai
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Get API key from environment variable
openai.api_key = os.getenv("OPENAI_API_KEY")

# Make the API call
response = openai.chat.completions.create(
    model="gpt-3.5-turbo",
    messages=[
        {"role": "system", "content": "You are a helpful assistant."},
        {"role": "user", "content": "Which is the best 3 day workout split?"}
    ],
    temperature=0.9,
    max_tokens=50
)

# Print the response
print(response.choices[0].message.content)